// src/types/animal.ts
// Generated by https://quicktype.io

export interface Animal {
  entries: Entry[]
  meta: AnimalMeta
}

export interface Entry {
  meta: EntryMeta
  fields: Fields
}

export interface Fields {
  image: Image
}

export interface Image {
  url: string
  tags: any[]
  uuid: string
  title: string
  alt_text: null
  description: null
  content_type: string
}

export interface Card extends Image {
  matched: boolean
}

export enum ContentType {
  ImageJPEG = 'image/jpeg'
}

export interface EntryMeta {
  name: string
  slug: string
  tags: any[]
  type: Type
  uuid: string
  space: Space
  author: Author
  locale: Locale
  excerpt: string
  private: boolean
  targets: any[]
  category: null
  created_at: string
  updated_at: string
  published_at: string
  version_type: VersionType
  category_name: null
  category_slug: null
  unpublished_at: null
  available_locales: Locale[]
}

export interface Author {}

export enum Locale {
  Es = 'es'
}

export enum Space {
  Animals = 'animals'
}

export enum Type {
  Game = 'game'
}

export enum VersionType {
  Current = 'current'
}

export interface AnimalMeta {
  total_entries: number
  per_page: number
  current_page: number
  total_pages: number
}